Fixed production build
Added new shortcodes for adding dynamic url to static folder depending on the production or dev build. Reworked postcss.config.js to work with stats file and fixed urls of the assets to link to the themes folder instead of the documentation one if the postcss.config.js is not overwritten.